+++ CHANGELOG.md
@@ -2,10 +2,11 @@
 * Database path is hadrcoded
 * Added command-line argument to add hosts to database
 * Added command-line argument to list entries from database
+* Added "Shutdown" event processing
 * Fixed pointer error in dba_free()
 
 ## Version 0.2.1
-* SQLite replaced with BerkleyDB
+* SQLite storage replaced with BerkleyDB
 
 ## Version 0.1d
 * Added configurable internal network